+++ packages/at_client/lib/src/crypto/crypto.dart
@@ -4,6 +4,7 @@ import 'package:at_client/src/crypto/nskey/ck_manager.dart';
 import 'package:at_client/src/crypto/nskey/content_key.dart';
 import 'package:at_client/src/crypto/nskey/nskey_key_ring.dart';
 import 'package:at_client/src/crypto/nskey/nskey_provider.dart';
+import 'package:at_client/src/crypto/nskey/nskey_resolver.dart';
 import 'package:at_client/src/crypto/nskey/symmetric_aes_gcm_provider.dart';
 import 'package:at_commons/at_commons.dart';
 
@@ -16,6 +17,7 @@ export 'package:at_client/src/crypto/nskey/ck_manager.dart';
 export 'package:at_client/src/crypto/nskey/content_key.dart';
 export 'package:at_client/src/crypto/nskey/nskey_key_ring.dart';
 export 'package:at_client/src/crypto/nskey/nskey_provider.dart';
+export 'package:at_client/src/crypto/nskey/nskey_resolver.dart';
 export 'package:at_client/src/crypto/nskey/published_nskey_key_ring.dart';
 export 'package:at_client/src/crypto/nskey/symmetric_aes_gcm_provider.dart';